backend/requirements.txt
fastapi
uvicorn
langgraph
langfuse
torch
transformers
pillow
pytesseract
pandas
